Minor fixes linked to deprecated types and dependencies (#122)
* Replace deprecated np types

* Comment python dependencies from package.xml, as they no longer work

+++ rt_bene_model_training/pytorch/rtbene_dataset.py
@@ -65,7 +65,7 @@ class RTBENEH5Dataset(data.Dataset):
         with h5py.File(self._h5_file, mode="r") as h5_file:
             left_img = h5_file[sample[0] + "/left"][sample[1]][()][0]
             right_img = h5_file[sample[0] + "/right"][sample[1]][()][0]
-            label = h5_file[sample[0] + "/label"][()].astype(np.float32)
+            label = h5_file[sample[0] + "/label"][()].astype(float)
 
             # Load data and get label
             transformed_left_img = self._transform(Image.fromarray(left_img, 'RGB'))
